requirements.txt
django==2.2.10
django-redis
djangorestframework
django-filter
django-cors-headers
django-imagekit
django-debug-toolbar
django-rest-swagger
PyJWT
qiniu
qrcode
openpyxl
xlrd
xlwt
requests
celery
itsdangerous
elasticsearch
PyMySQL
pillow
boto3
ujson
flower
django-celery-results
alipay-sdk-python